pyproject.toml
[tool.poetry]
name = "graphql_project"
version = "0.1.0"
description = ""
authors = ["VictorAugustoBarros <victor.augustobarros@gmail.com>"]
packages = [
  { include = "connections", from = "app" },
  { include = "graphql_query", from = "app" },
  { include = "models", from = "app" },
  { include = "routes", from = "app" },
  { include = "utils", from = "app" },
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.81.0"
strawberry-graphql = "^0.127.4"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
black = "^22.8.0"
commitizen = "^2.32.2"
pre-commit = "^2.20.0"
pydocstyle = "^6.1.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
api = 'app.main:api.run'


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
